requirements.txt
numpy
scipy
matplotlib
pillow
rawpy
